# backend/public/.well-known/commerce-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Agentrix Commerce Skill API",
    "description": "Unified commerce protocol for AI agents: payment, revenue splitting, budget pools, and marketplace.",
    "version": "1.0.0",
    "contact": { "url": "https://www.agentrix.top" }
  },
  "servers": [
    { "url": "https://www.agentrix.top/api" }
  ],
  "paths": {
...
    "/commerce/publish": {
      "post": {
        "operationId": "publishToMarketplace",
        "summary": "Publish a skill, product, service, or task to Agentrix Marketplace (www.agentrix.top)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "category": { "type": "string" }, "pricingType": { "type": "string", "enum": ["free", "per_call", "subscription", "one_time"] }, "pricePerCall": { "type": "number" } }, "required": ["name", "description"] } } }
        },
        "responses": { "201": { "description": "Published item" } }
      }
    },
    "/commerce/fees/calculate": {
      "post": {
        "operationId": "calculateFees",
        "summary": "Calculate fees for a transaction. Crypto direct: 0%, OnRamp: +0.1%, OffRamp: +0.1%, Split: 0.3% (min 0.1 USDC)",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "amount": { "type": "number" }, "includeOnramp": { "type": "boolean" }, "includeOfframp": { "type": "boolean" }, "includeSplit": { "type": "boolean" } } } } }
        },
        "responses": { "200": { "description": "Fee calculation result" } }
      }
    },
...
  }
}